Refactoring remaining commands
