+++ cli/src/__tests__/commands/fixtures/invalid.yaml
