+++ cli/src/__tests__/commands/fixtures/valid.yaml
