+++ cli/src/__tests__/commands/init/init.test.ts
@@ -1,9 +1,9 @@
 import { describe, it, expect, beforeEach, jest } from "@jest/globals";
 import { Command } from "commander";
-import { initCommand } from "../../commands/init";
-import { DefaultInitService } from "../../services/init.service";
+import { initCommand } from "../../../commands/init/init";
+import { DefaultInitService } from "../../../commands/init/init.service";
 
-jest.mock("../../services/init.service");
+jest.mock("../../../commands/init/init.service");
 
 describe("initCommand", () => {
   let program: Command;
