+++ cli/src/commands/compile/compile.service.ts
@@ -1,8 +1,7 @@
-import { CompileService } from "./interfaces";
 import { spawn } from "child_process";
-import { tspBinPath } from "../utils/typespec";
+import { tspBinPath } from "../../utils/typespec";
 
-export class DefaultCompileService implements CompileService {
+export class DefaultCompileService {
   async compile(specPath: string): Promise<void> {
     return new Promise((resolve, reject) => {
       const process = spawn("node", [tspBinPath, "compile", specPath], {
