+++ cli/src/commands/preview/preview.ts
@@ -1,6 +1,6 @@
 import { Command } from "commander";
-import { DefaultPreviewService } from "../services/preview.service";
-import { PreviewArgsSchema } from "../types/command-args";
+import { DefaultPreviewService } from "./preview.service";
+import { PreviewArgsSchema } from "../../types/command-args";
 
 export function previewCommand(program: Command) {
   const previewService = new DefaultPreviewService();
