+++ cli/src/index.ts
@@ -1,10 +1,10 @@
 #!/usr/bin/env node
 
 import { program } from "commander";
-import { initCommand } from "./commands/init";
+import { initCommand } from "./commands/init/init";
 import { previewCommand } from "./commands/preview";
 import { checkCommand } from "./commands/check/check";
-import { compileCommand } from "./commands/compile";
+import { compileCommand } from "./commands/compile/compile";
 
 program.name("cg").description("CommonGrants CLI tools").version("0.1.0");
